fixed type

+++ 0x00-hello_world/6-size.c
@@ -9,8 +9,8 @@ int main(void)
 {
 	printf("Size of a char: %ld bytes(s)\n", sizeof(char));
 	printf("Size of an int: %ld bytes(s)\n", sizeof(int));
-	printf("Size of long int: %ld bytes(s)\n", sizeof(int));
-	printf("Size of long long int: %lu bytes(s)\n", sizeof(long));
+	printf("Size of long int: %ld bytes(s)\n", sizeof(long int));
+	printf("Size of long long int: %ld bytes(s)\n", sizeof(long long int));
 	printf("Size of float: %ld bytes(s)\n", sizeof(float));
 
 	return (0);
